libpressio version 0.88.0
Major Changes

+ **Breaking Change** pressio_thread_safety and pressio_dtype can now be
  stored in pressio_option structures.  This change improves ergonomics
  and typesafety for options using these types.  Unfortunately, this
  breaks the API for users of pressio:thread_safe.
+ normalize_dims now allows the user to specify the size and fill for
  dimension normalization.  Once fully supported, all compressors can
  trivially support higher dimensions by concatenation.

Minor Changes

+ Python bindings now load 3rd party compressors by default

Bug Fixes

+ many_independent_threaded previously didn't check when updating its
  array of metrics results if the previous result were empty.  Now it
  does so meaningful results are returned.
+ Missing header in data.h was added

+++ src/plugins/compressors/blosc_plugin.cc
@@ -41,7 +41,7 @@ class blosc_plugin: public libpressio_compressor_plugin {
 
     struct pressio_options get_configuration_impl() const override {
       struct pressio_options options;
-      set(options, "pressio:thread_safe", static_cast<int32_t>(pressio_thread_safety_multiple));
+      set(options, "pressio:thread_safe", pressio_thread_safety_multiple);
       set(options, "pressio:stability", "stable");
       std::vector<std::string> compiled_compressors;
       std::string s = blosc_list_compressors();
