+++ src/plugins/compressors/delta_encoding.cc
@@ -63,7 +63,7 @@ y[i] = x[i] - x[i-1];
   pressio_options get_configuration_impl() const override {
     pressio_options opts;
     opts.copy_from(meta->get_configuration());
-    set(opts, "pressio:thread_safe", static_cast<int32_t>(get_threadsafe(*meta)));
+    set(opts, "pressio:thread_safe", get_threadsafe(*meta));
     set(opts, "pressio:stability", "experimental");
     return opts;
   }
