+++ src/plugins/compressors/mgard_plugin.cc
@@ -128,7 +128,7 @@ class mgard_plugin: public libpressio_compressor_plugin {
 
   struct pressio_options get_configuration_impl() const override {
     struct pressio_options options;
-    set(options, "pressio:thread_safe", static_cast<int32_t>(pressio_thread_safety_single));
+    set(options, "pressio:thread_safe", pressio_thread_safety_single);
     set(options, "pressio:stability", "experimental");
     return options;
   }
