+++ src/plugins/compressors/noop.cc
@@ -17,7 +17,7 @@ class noop_compressor_plugin: public libpressio_compressor_plugin {
 
   struct pressio_options get_configuration_impl() const override {
     pressio_options options;
-    set(options, "pressio:thread_safe", static_cast<int32_t>(pressio_thread_safety_multiple));
+    set(options, "pressio:thread_safe", pressio_thread_safety_multiple);
     set(options, "pressio:stability", "stable");
     return options;
   }
