+++ src/plugins/compressors/sz_threadsafe_plugin.cc
@@ -54,7 +54,7 @@ class sz_threadsafe_plugin: public libpressio_compressor_plugin {
 
   struct pressio_options get_configuration_impl() const override {
     struct pressio_options options;
-    set(options, "pressio:thread_safe", static_cast<int32_t>(pressio_thread_safety_multiple));
+    set(options, "pressio:thread_safe", pressio_thread_safety_multiple);
     set(options, "pressio:stability", "stable");
 #ifdef HAVE_RANDOMACCESS
     set(options, "sz_threadsafe:random_access_enabled", 1u);
